+++ lib/queries/one/aggregatable_table.ts
@@ -1,6 +1,5 @@
 export type AggregatableColumn<F> = {
     avg(): F
-    count(): F
     max(): F
     min(): F
     sum(): F